+++ app/api/v1/routers.py
@@ -24,6 +24,29 @@ def get_airports(db: Session = Depends(get_db)) -> list[Airport]:
     return get_airports_from_db(db=db)
 
 
+@router.get('/search')
+def get_flight_search_result():
+    #--- get ajax POST data
+    time_limit = requests.json["time_limit"]
+    expense_limit = requests.json["expense_limit"]
+    current_lat = requests.json["current_lat"]
+    current_lng = requests.json["current_lng"]
+    print("main.py ajax POST data - time_limit: " + time_limit)
+    print("main.py ajax POST data - expense_limit: " + expense_limit)
+    print("main.py ajax POST data - current_lat: " + current_lat)
+    print("main.py ajax POST data - current_lng: " + current_lng)
+
+    #--- search and get near airport from MySQL (airport table)
+    near_airport_IATA = get_near_airport(current_lat,current_lng)
+    print("main.py get values - near_airport_IATA: " + near_airport_IATA)
+
+    #--- search and get reachable location (airport and country) from skyscanner api
+    #--- exclude if time and travel expenses exceed the user input parameter
+    #--- select a country at random
+    destination = get_destination_from_skyscanner_by_random(near_airport_IATA,time_limit,expense_limit)
+    return destination
+
+
 @router.post('/shuffle')
 def get_random_country():
     result = get_country()
